"I PROVIDED NOTICE AT ALL RELEVANT TIMES UPON NOTICE, AND THOSE WHO HAVE AIDED AND ...
"I PROVIDED NOTICE AT ALL RELEVANT TIMES UPON NOTICE, AND THOSE WHO HAVE AIDED AND
ABETTED TO THE NON DISCLOSURE OF A FAILURE AND OMISSION TO DISCLOSE AN UNREGISTERED
SECURITY (THE LOAN 50074) ANNEXED IN THE NYSCEF MATTER 153974/2020 HAVE CAUSED OVER
(852,029,489.38) IN DAMAGES TO THE INVESTORS - NOTWITHSTANDING THE FILING BY MR. DAVID
MOORE - david.moore.ct95@statefarm.com - AND THE TCR WHICH I FILED DURING THE MERGER PERIOD
WAS ALSO FURTHER OBSTRUCTED BY THE GUIDANCE COUNSELORS AT MY SCHOOL ON NOVEMBER
16TH, 2021 - I NOTIFIED ONE OF THREE ""PROMOTING"" BROKERS AND ALSO THE SEC OF THE MATERIAL
FACTS AGAIN IN NYSCEF MATTER 153974/2021."			
"MY EMAIL WAS TURNED OFF -- BY A GUIDANCE COUNSELOR AT COLUMBIA UNIVERSITY AND DURING
A MATERIAL AND TIME-SENSITIVE PERIOD IN TIME... NOTWITHSTANDING THE ASSISTANCE AND
COMMUNICATIONS AS INDEXED IN THE MATTER OF 153974/2020 ---- WHERE IGNORANCE AND A
LACK OF INTELLIGENCE/CANDOR/DISRESPECT HAVE CAUSED, IN PART, TO THE AID AND ABETTING
OF TAX EVASION AS SEEN IN THE TAX RECORDS I ALSO ANNEXED IN A PUBLIC REPOSITORY - THE
NYS TAX RECORDS OF THE SULLIVAN PROPERTIES L.P. SHOW THIS FOR THE LOAN (UNDISCLOSED)
WAS NOT COVERED UNDER OMISSIONS BY THE PRINCIPALS - AT ANY POINT IN TIME - WOULD DO
ANYTHING TO PREVENT THIS CERTAIN DELUGE AND THE UNDERLYING RISKS ARE NOW CLEAR - IN
A FOUR MONTH PERIOD A TOTAL OF (852,029,489.38) IN ACCOUNTABLE LOSSES FOR THE TICKERS...
NOTWITHSTANDING A LOSS BY WAY OF TAXES AVOIDED TO BY THE ZUCKER FAMILY AND SULLIVAN
PROPERTIES L.P. IN THE NYSCEF MATTER 153974/2020."			

		
			
..2022-04-28---TCRReport 16511-049-089-843		
FILED IN FURTHERANCE/ SUPPLEMENT TO TCR  DOCKET	16491-117-831-823		
Fraudulent investment scheme, such as a Ponzi scheme or the promise of high-yield returns			
SUBJECTS	TYPE	IDENTIFIER	NAME
Subject # 1	Firm	Unknown	STATE FARM MUTUAL AUTOMOBILE INSURANCE COMPANY

		
			
"SUBSCRIBERS ADDED WITHOUT NOTICE OR DISCLOSURE OF ILLEGAL RENTS AND LEASES
TRANSFERRED AND A LETTER OF CREDIT FOR $6,000,000.00 is not reported as a level 3 investment and
undisclosed New York Supreme Case 153974/2020 where I annexed the Loan Dockets (see also Docket 420)"			
"This ZUCKER FAMILY WILL DO ANYTHING TO AVOID JAIL FOR THE $300MILLION THEY OWE IN BACKTAXES...
SEE ALSO DOCKET 399. I emailed a bunch of information to make this as complicated as possible,
after trying to help them to CANCEL THE LOAN as annexed in DOCKET 420 (was during the 90-day period,
in lieu they decided to keep the premiums and interest - despite knowing this information: https://github.com/
users/BSCPGROUPHOLDINGSLLC/projects/1#card-79888540) which layman's terms, they loaned $6 million
dollars and now STATE FARM OWNS THE RIGHTS to the taxes evaded in the ASSETS which guarantee the"			
"leases and RENTS are taxed at %10 of the NOPV as of current. The 6 buildings are being taxed at $22 million
(combined) when they are closer to $22 Million EACH (hence back taxed for 10-years prior, and the carried interest
and penalties is the greater of $300,000,000.00 in fines, and the property values ON TOP, were reported to the
NYC Finance Register as 10% YoY gain as referenced in the public domain (for consideration by the criminally
insane). For ease of reference I have included the case matter where I annexed there are NO CERTIFICATES OF
OCCUPANCY for them to collect LEGAL rents which were reported to the NYC DEPT. OF FINANCE REGISTER
with a YoY growth of 10%. Under any normal circumstances would be FLAGGED by a Manhattan Bank, in
any zipcode, was avoided in the compliant in CHIEF at all costs to avoid prosecution in a CRIMINAL or CIVIL
COURT OF LAW. Earlier I attempted to contact parties in good faith, and their brokers took it upon themselves
to use the information to their benefit without disseminating the material fact of a CERTAIN DELUGE and
immediately upon NOTICE in DECEMBER of 2021, the family of FUNDS who operated under CIK Filer 93715
appeared to have been ""clothes-lined"" despite the ignorance of my filing on November 13th, 2021 (which was
also distributed to the C.16 fiduciaries) and in a posthaste negligence to those facts decided to ""merge"" and
maintain the same compliance officer at State Farm VP. Management Corp and incurred additional losses since
the 4th of April, 2022 [ TCRReport-16491-117-831-823 ] in excess of 1.2 billion dollars (US) - continues to operate
without any public information of these MATERIAL DOCKETS annexed in 153974/2020 which also demonstrate
CERTAIN tax evasion without any doubt. #GOCARDS ""https://github.com/users/BSCPGROUPHOLDINGSLLC/
projects/1#column-18309490"" was also aided and abetted by its attorneys, notwithstanding The Wilson, Elser
(www.wilsonelser.com) and Dicker law firm --- BE IT REMEMBERED."			
"No response since I annexed the TAX EVASION and docketed no certificates of occupancy and a $6 million dollar
line of credit extended by State Farm located at one state farm plaza, in bloomington, Illinois, 61710.

****https://iapps.courts.state.ny.us/nyscef/ViewDocument?docIndex=19MVPFXy0G0QvnmRLGpYIQ==<br>

CERTIFICATION			
"###Pursuant to Section 906, of the Sarbanes-Oxley Act of 2002 and section 63, Title 17.: VIOLATION ### 
***""https://www.sec.gov/Archives/edgar/data/0000093715/000119312520200810/d913497dex99906cert.htm<br>

 HAVE LOANED $6,000,000.00 of CASH for a purported series of RENTS, SECURITY DEPOSITS, 
	AND LEASES WHICH ARE ALL UNLAWFULLY CUSTODIED PER THE FEDERAL DEPOSIT INSURANCE CORPORATION 
	
		in the UNITED STATES OF AMERICA --- are now ""SELF-ADMINISTERED"" and a ""SELF-CUSTODIAN""
		by the new IMPROVED Investment Adviser who is also Audited by THREE PUBLICLY TRADED AUDIT FIRMS 
		under the sub-adviser which previously serviced there Transactions, has been promoted 
		in the NEW CIK 1516523. ""THE SURVIVING ENTITY"" despite my NOTICE and filing 
		on November 13th, 2021 during the ""PERIOD OF MERGER"""			
			
			
			

https://www.fbi.gov/investigate/white-collar-crime
			
https://www.fbi.gov/file-repository/criminal-and-epidemiological-investigation-handbook.pdf
			
https://github.com/BSCPGROUPHOLDINGSLLC/ELSER-AND-DICKER/blob/BSCPGROUPHOLDINGSLLC-EMAIL-DOCKETS/ms-pwc-dicker
		
		
The Securities and Exchange Commission on Monday brought charges against 11 individuals for allegedly bilking investors out of more than $300 million in a crypto pyramid and Ponzi scheme. Regulators allege the scheme started in at least January 2020.Aug 1, 2022
https://www.dni.gov/